requirements.txt
requests
redis
python-dotenv
pcloud
